+++ platform/viewer/cypress/integration/common/OHIFUserPreferences.spec.js
@@ -30,9 +30,8 @@ describe('OHIF User Preferences', () => {
     });
 
     it('checks translation by selecting Spanish language', function() {
-      cy.get('@userPreferencesGeneralTab')
-        .click()
-        .should('have.class', 'active');
+      cy.changePreferencesTab('@userPreferencesGeneralTab');
+      cy.get('@userPreferencesGeneralTab').should('have.class', 'active');
 
       // Language dropdown should be displayed
       cy.get('#language-select').should('be.visible');
@@ -65,14 +64,14 @@ describe('OHIF User Preferences', () => {
       cy.get('[data-cy="options-menu"]').click();
     });
 
-    it('checks if user can cancel the language selection and application will be in English', function() {
+    it('checks if user can cancel the language selection and application will be in "English (USA)"', function() {
       // Set language to English and save
-      cy.setLanguage('English');
+      cy.setLanguage('English (USA)');
 
       // Set language to Spanish and cancel
       cy.setLanguage('Spanish', false);
 
-      // Header should be kept in English
+      // Header should be kept in "English (USA)"
       cy.get('.research-use')
         .scrollIntoView()
         .should('have.text', 'INVESTIGATIONAL USE ONLY');
@@ -93,7 +92,7 @@ describe('OHIF User Preferences', () => {
       cy.get('[data-cy="options-menu"]').click();
     });
 
-    it('checks if user can restore to default the language selection and application will be in English', function() {
+    it('checks if user can restore to default the language selection and application will be in "English (USA)"', function() {
       // Set language to Spanish
       cy.setLanguage('Spanish');
 
@@ -101,7 +100,7 @@ describe('OHIF User Preferences', () => {
       cy.openPreferences();
 
       // Go to general tab
-      cy.get('@userPreferencesGeneralTab').click();
+      cy.changePreferencesTab('@userPreferencesGeneralTab');
 
       cy.get('@restoreBtn')
         .scrollIntoView()
@@ -112,12 +111,12 @@ describe('OHIF User Preferences', () => {
         .scrollIntoView()
         .click();
 
-      // Header should be in English
+      // Header should be in "English (USA)"
       cy.get('.research-use')
         .scrollIntoView()
         .should('have.text', 'INVESTIGATIONAL USE ONLY');
 
-      // Options menu should be in English
+      // Options menu should be in "English (USA)"
       cy.get('[data-cy="options-menu"]')
         .should('have.text', 'Options')
         .click();
@@ -135,9 +134,8 @@ describe('OHIF User Preferences', () => {
 
     it('checks if Preferences set in Study List Page will be consistent on Viewer Page', function() {
       // Go go hotkeys tab
-      cy.get('@userPreferencesHotkeysTab')
-        .click()
-        .should('have.class', 'active');
+      cy.changePreferencesTab('@userPreferencesHotkeysTab');
+      cy.get('@userPreferencesHotkeysTab').should('have.class', 'active');
 
       // Set new hotkey for 'Rotate Right' function
       cy.setNewHotkeyShortcutOnUserPreferencesModal('Rotate Right', '{shift}Q');
@@ -150,7 +148,7 @@ describe('OHIF User Preferences', () => {
       cy.openPreferences();
 
       // Go to General tab
-      cy.get('@userPreferencesGeneralTab').click();
+      cy.changePreferencesTab('@userPreferencesGeneralTab');
 
       // Set language to Spanish
       cy.setLanguage('Spanish');
@@ -221,9 +219,8 @@ describe('OHIF User Preferences', () => {
     });
 
     it('checks translation by selecting Spanish language', function() {
-      cy.get('@userPreferencesGeneralTab')
-        .click()
-        .should('have.class', 'active');
+      cy.changePreferencesTab('@userPreferencesGeneralTab');
+      cy.get('@userPreferencesGeneralTab').should('have.class', 'active');
 
       // Language dropdown should be displayed
       cy.get('#language-select').should('be.visible');
@@ -256,14 +253,14 @@ describe('OHIF User Preferences', () => {
       cy.get('[data-cy="options-menu"]').click();
     });
 
-    it('checks if user can cancel the language selection and application will be in English', function() {
+    it('checks if user can cancel the language selection and application will be in "English (USA)"', function() {
       // Set language to English and save
-      cy.setLanguage('English');
+      cy.setLanguage('English (USA)');
 
       // Set language to Spanish and cancel
       cy.setLanguage('Spanish', false);
 
-      // Header should be kept in English
+      // Header should be kept in "English (USA)"
       cy.get('.research-use')
         .scrollIntoView()
         .should('have.text', 'INVESTIGATIONAL USE ONLY');
@@ -283,10 +280,9 @@ describe('OHIF User Preferences', () => {
       cy.get('[data-cy="options-menu"]').click();
     });
 
-    it('checks if user can restore to default the language selection and application will be in English', function() {
-      cy.get('@userPreferencesGeneralTab')
-        .click()
-        .should('have.class', 'active');
+    it('checks if user can restore to default the language selection and application will be in "English (USA)', function() {
+      cy.changePreferencesTab('@userPreferencesGeneralTab');
+      cy.get('@userPreferencesGeneralTab').should('have.class', 'active');
 
       // Language dropdown should be displayed
       cy.get('#language-select').should('be.visible');
@@ -298,7 +294,7 @@ describe('OHIF User Preferences', () => {
       cy.openPreferences();
 
       // Go to general tab
-      cy.get('@userPreferencesGeneralTab').click();
+      cy.changePreferencesTab('@userPreferencesGeneralTab');
 
       cy.get('@restoreBtn')
         .scrollIntoView()
@@ -308,12 +304,12 @@ describe('OHIF User Preferences', () => {
         .scrollIntoView()
         .click();
 
-      // Header should be in English
+      // Header should be in "English (USA)""
       cy.get('.research-use')
         .scrollIntoView()
         .should('have.text', 'INVESTIGATIONAL USE ONLY');
 
-      // Options menu should be in English
+      // Options menu should be in "English (USA)"
       cy.get('[data-cy="options-menu"]')
         .should('have.text', 'Options')
         .click();
@@ -330,9 +326,8 @@ describe('OHIF User Preferences', () => {
 
     it('checks new hotkeys for "Rotate Right" and "Rotate Left"', function() {
       // Go go hotkeys tab
-      cy.get('@userPreferencesHotkeysTab')
-        .click()
-        .should('have.class', 'active');
+      cy.changePreferencesTab('@userPreferencesHotkeysTab');
+      cy.get('@userPreferencesHotkeysTab').should('have.class', 'active');
 
       // Set new hotkey for 'Rotate Right' function
       cy.setNewHotkeyShortcutOnUserPreferencesModal(
@@ -360,24 +355,17 @@ describe('OHIF User Preferences', () => {
     });
 
     it('checks new hotkeys for "Next" and "Previous" Image on Viewport', function() {
-      // Go go hotkeys tab
-      cy.get('@userPreferencesHotkeysTab')
-        .click()
-        .should('have.class', 'active');
-
-      // Set new hotkey for 'Next Image Viewport' function
+      // Update hotkeys for 'Next/Previous Viewport'
+      cy.changePreferencesTab('@userPreferencesHotkeysTab');
+      cy.get('@userPreferencesHotkeysTab').should('have.class', 'active');
       cy.setNewHotkeyShortcutOnUserPreferencesModal(
-        'Next Image Viewport',
+        'Next Viewport',
         '{shift}{rightarrow}'
       );
-
-      // Set new hotkey for 'Previous Image Viewport' function
       cy.setNewHotkeyShortcutOnUserPreferencesModal(
-        'Previous Image Viewport',
+        'Previous Viewport',
         '{shift}{leftarrow}'
       );
-
-      // Save new hotkeys
       cy.get('@saveBtn')
         .scrollIntoView()
         .click();
@@ -386,20 +374,25 @@ describe('OHIF User Preferences', () => {
       cy.setLayout(3, 1);
       cy.waitViewportImageLoading();
 
-      // Rotate Right and Invert colors on Viewport #1
-      cy.get('body').type('RI');
-      // Check that image was rotated
+      // Reset, Rotate Right and Invert colors on Viewport #1
+      cy.get('body').type(' ');
+      cy.get('body').type('r');
+      cy.get('body').type('i');
+
+      // Shift active viewport to next
+      // Reset, Rotate Left and Invert colors on Viewport #2
+      cy.get('body').type('{shift}{rightarrow}');
+      cy.get('body').type(' ');
+      cy.get('body').type('l');
+      cy.get('body').type('i');
+
+      // Verify 1st viewport was rotated
       cy.get('@viewportInfoMidTop').should('contains.text', 'R');
 
-      //Move to Next Viewport
-      cy.get('body').type('{shift}{rightarrow}');
-      // Rotate Left and Invert colors on Viewport #2
-      cy.get('body').type('LI');
-      // Get overlay information from viewport #2
+      // Verify 2nd viewport was rotated
       cy.get(
         ':nth-child(2) > .viewport-wrapper > .viewport-element > .ViewportOrientationMarkers.noselect > .top-mid.orientation-marker'
       ).as('viewport2InfoMidTop');
-      // Check that image was rotated
       cy.get('@viewport2InfoMidTop').should('contains.text', 'P');
 
       //Move to Previous Viewport
@@ -417,24 +410,18 @@ describe('OHIF User Preferences', () => {
 
     it('checks error message when duplicated hotkeys are inserted', function() {
       // Go go hotkeys tab
-      cy.get('@userPreferencesHotkeysTab').click();
+      cy.changePreferencesTab('@userPreferencesHotkeysTab');
 
       // Set duplicated hotkey for 'Rotate Right' function
-      cy.setNewHotkeyShortcutOnUserPreferencesModal(
-        'Rotate Right',
-        '{rightarrow}'
-      );
+      cy.setNewHotkeyShortcutOnUserPreferencesModal('Rotate Right', '{i}');
 
       // Check error message
-      cy.get('.HotKeysPreferences').within(() => {
+      cy.get('.HotkeysPreferences').within(() => {
         cy.contains('Rotate Right') // label we're looking for
           .parent()
-          .find('.errorMessage')
+          .find('.preferencesInputErrorMessage')
           .as('errorMsg')
-          .should(
-            'have.text',
-            '"Next Image Viewport" is already using the "right" shortcut.'
-          );
+          .should('have.text', '"Invert" is already using the "i" shortcut.');
       });
       //Cancel hotkeys
       cy.get('@cancelBtn')
@@ -444,16 +431,16 @@ describe('OHIF User Preferences', () => {
 
     it('checks error message when invalid hotkey is inserted', function() {
       // Go go hotkeys tab
-      cy.get('@userPreferencesHotkeysTab').click();
+      cy.changePreferencesTab('@userPreferencesHotkeysTab');
 
       // Set invalid hotkey for 'Rotate Right' function
       cy.setNewHotkeyShortcutOnUserPreferencesModal('Rotate Right', '{ctrl}Z');
 
       // Check error message
-      cy.get('.HotKeysPreferences').within(() => {
+      cy.get('.HotkeysPreferences').within(() => {
         cy.contains('Rotate Right') // label we're looking for
           .parent()
-          .find('.errorMessage')
+          .find('.preferencesInputErrorMessage')
           .as('errorMsg')
           .should('have.text', '"ctrl+z" shortcut combination is not allowed');
       });
@@ -466,15 +453,15 @@ describe('OHIF User Preferences', () => {
 
     it('checks error message when only modifier keys are inserted', function() {
       // Go go hotkeys tab
-      cy.get('@userPreferencesHotkeysTab').click();
+      cy.changePreferencesTab('@userPreferencesHotkeysTab');
 
       // Set invalid modifier key: ctrl
       cy.setNewHotkeyShortcutOnUserPreferencesModal('Zoom Out', '{ctrl}');
       // Check error message
-      cy.get('.HotKeysPreferences').within(() => {
+      cy.get('.HotkeysPreferences').within(() => {
         cy.contains('Zoom Out') // label we're looking for
           .parent()
-          .find('.errorMessage')
+          .find('.preferencesInputErrorMessage')
           .as('errorMsg')
           .should(
             'have.text',
@@ -506,7 +493,7 @@ describe('OHIF User Preferences', () => {
 
     it('checks if user can cancel changes made on User Preferences Hotkeys tab', function() {
       // Go go hotkeys tab
-      cy.get('@userPreferencesHotkeysTab').click();
+      cy.changePreferencesTab('@userPreferencesHotkeysTab');
 
       // Set new hotkey for 'Rotate Right' function
       cy.setNewHotkeyShortcutOnUserPreferencesModal(
@@ -523,7 +510,7 @@ describe('OHIF User Preferences', () => {
       cy.openPreferences();
 
       //Check that hotkey for 'Rotate Right' function was not changed
-      cy.get('.HotKeysPreferences').within(() => {
+      cy.get('.HotkeysPreferences').within(() => {
         cy.contains('Rotate Right') // label we're looking for
           .parent()
           .find('input')
@@ -534,7 +521,7 @@ describe('OHIF User Preferences', () => {
 
     it('checks if user can reset to default values on User Preferences Hotkeys tab', function() {
       // Go go hotkeys tab
-      cy.get('@userPreferencesHotkeysTab').click();
+      cy.changePreferencesTab('@userPreferencesHotkeysTab');
 
       // Set new hotkey for 'Rotate Right' function
       cy.setNewHotkeyShortcutOnUserPreferencesModal(
@@ -559,7 +546,7 @@ describe('OHIF User Preferences', () => {
       cy.openPreferences();
 
       //Check that hotkey for 'Rotate Right' function was not changed
-      cy.get('.HotKeysPreferences').within(() => {
+      cy.get('.HotkeysPreferences').within(() => {
         cy.contains('Rotate Right') // label we're looking for
           .parent()
           .find('input')