+++ platform/viewer/src/appExtensions/GenericViewerCommands/index.js
@@ -2,7 +2,7 @@ import commandsModule from './commandsModule.js';
 
 export default {
   id: 'generic-viewer-commands',
-  getCommandsModule() {
-    return commandsModule;
+  getCommandsModule({ commandsManager }) {
+    return commandsModule({ commandsManager });
   },
 };